refactor: split LiveDashboard into sub-components
