--- src/components/professor/LiveDashboard.tsx
@@ -1,501 +1,0 @@
-"use client";
-
-import { useEffect, useState } from "react";
-import { Session, Question, StudentResponse } from "@/lib/types";
-import { collection, query, onSnapshot, doc, updateDoc } from "firebase/firestore";
-import { db, auth } from "@/lib/firebase/client";
-import { Card, CardHeader, CardTitle, CardContent } from "@/components/ui/card";
-import { Button } from "@/components/ui/button";
-import { BarChart, Bar, XAxis, YAxis, Tooltip, ResponsiveContainer, Cell } from "recharts";
-import { motion, AnimatePresence } from "framer-motion";
-import { Users, StopCircle, QrCode, Loader2, Sparkles, Bot, Plus, Play, EyeOff, Trash2, ArrowLeft } from "lucide-react";
-import { toast } from "sonner";
-import { Dialog, DialogContent, DialogTrigger } from "@/components/ui/dialog";
-import Link from "next/link";
-import { SessionQR } from "@/components/professor/SessionQR";
-import { StarRating } from "@/components/ui/StarRating";
-import { Input } from "@/components/ui/input";
-import { Label } from "@/components/ui/label";
-import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from "@/components/ui/select";
-import { isDemoSession, updateLocalSession } from "@/lib/storage";
-
-export function LiveDashboard({ session }: { session: Session }) {
-    const [responses, setResponses] = useState<StudentResponse[]>([]);
-
-    useEffect(() => {
-        if (session.id && session.id.startsWith("local_")) {
-            return;
-        }
-
-        const q = query(collection(db, "sessions", session.id!, "responses"));
-        const unsub = onSnapshot(q, (snapshot) => {
-            const list = snapshot.docs.map(d => d.data() as StudentResponse);
-            setResponses(list);
-        }, (err) => {
-            console.error("LiveDashboard Error:", err);
-        });
-        return () => unsub();
-    }, [session.code, session.id]);
-
-    const [analyzing, setAnalyzing] = useState(false);
-    const [isAddOpen, setIsAddOpen] = useState(false);
-    const [newQText, setNewQText] = useState("");
-    const [newQType, setNewQType] = useState<"short_text" | "multiple_choice" | "file_upload" | "rating">("short_text");
-    const [newQOptions, setNewQOptions] = useState<string[]>(["Option 1", "Option 2"]);
-
-    const [analyzingIds, setAnalyzingIds] = useState<Set<string>>(new Set());
-
-
-
-    const handleAddQuestion = async () => {
-        if (!newQText) return;
-
-        const newQ: Question = {
-            id: Math.random().toString(36).substr(2, 9),
-            text: newQText,
-            type: newQType,
-            options: newQType === "multiple_choice" ? newQOptions : undefined,
-            isActive: true
-        };
-
-        const updatedQuestions = [...(session.questions || []), newQ];
-
-        try {
-            if (!session.id?.startsWith("local_")) {
-                const token = await auth.currentUser?.getIdToken();
-                await fetch(`/api/sessions/${session.id}/update`, {
-                    method: "POST",
-                    headers: { "Content-Type": "application/json", "Authorization": `Bearer ${token}` },
-                    body: JSON.stringify({ questions: updatedQuestions })
-                });
-            }
-            if (isDemoSession(session)) {
-                updateLocalSession(session.id!, { questions: updatedQuestions });
-                window.location.reload();
-            }
-
-            setIsAddOpen(false);
-            setNewQText("");
-            setNewQOptions(["Option 1", "Option 2"]);
-            toast.success("Question added!");
-        } catch (e) {
-            console.error(e);
-            toast.error("Failed to add question");
-        }
-    };
-
-    const toggleActiveQuestion = async (qId: string) => {
-        const currentActiveIds = new Set(session.activeQuestionIds || (session.activeQuestionId ? [session.activeQuestionId] : []));
-
-        let isNowActive = false;
-        if (currentActiveIds.has(qId)) {
-            currentActiveIds.delete(qId);
-            isNowActive = false;
-        } else {
-            currentActiveIds.add(qId);
-            isNowActive = true;
-        }
-
-        const newActiveIds = Array.from(currentActiveIds);
-
-        try {
-            if (!session.id?.startsWith("local_")) {
-                const token = await auth.currentUser?.getIdToken();
-                await fetch(`/api/sessions/${session.id}/update`, {
-                    method: "POST",
-                    headers: { "Content-Type": "application/json", "Authorization": `Bearer ${token}` },
-                    body: JSON.stringify({
-                        activeQuestionIds: newActiveIds,
-                        activeQuestionId: newActiveIds.length > 0 ? newActiveIds[newActiveIds.length - 1] : null // Keep legacy field partially in sync for safety
-                    })
-                });
-            } else if (isDemoSession(session)) {
-                updateLocalSession(session.id!, {
-                    activeQuestionIds: newActiveIds,
-                    activeQuestionId: newActiveIds.length > 0 ? newActiveIds[newActiveIds.length - 1] : null
-                });
-                window.location.reload();
-            }
-            if (isNowActive) {
-                toast.success("Question is NOW LIVE");
-            } else {
-                toast.success("Question stopped");
-            }
-        } catch (e) {
-            toast.error("Failed to update status");
-        }
-    };
-
-    const deleteQuestion = async (qId: string) => {
-        const updatedQuestions = session.questions.filter(q => q.id !== qId);
-        try {
-            if (!session.id?.startsWith("local_")) {
-                const token = await auth.currentUser?.getIdToken();
-                await fetch(`/api/sessions/${session.id}/update`, {
-                    method: "POST",
-                    headers: { "Content-Type": "application/json", "Authorization": `Bearer ${token}` },
-                    body: JSON.stringify({ questions: updatedQuestions })
-                });
-            } else if (isDemoSession(session)) {
-                updateLocalSession(session.id!, { questions: updatedQuestions });
-                window.location.reload();
-            }
-            toast.success("Question deleted");
-        } catch (e) {
-            toast.error("Failed to delete");
-        }
-    }
-
-    const closeSession = async () => {
-        setAnalyzing(true);
-        const analysisResults: Record<string, any> = {};
-
-        // Global synthesis is now handled in the SynthesisView manually after session close.
-        // We just close the session here.
-
-        try {
-            if (!session.id?.startsWith("local_")) {
-                const token = await auth.currentUser?.getIdToken();
-                await fetch(`/api/sessions/${session.id}/update`, {
-                    method: "POST",
-                    headers: { "Content-Type": "application/json", "Authorization": `Bearer ${token}` },
-                    body: JSON.stringify({
-                        status: "CLOSED",
-                        ...(Object.keys(analysisResults).length > 0 ? { analysis: analysisResults } : {})
-                    })
-                });
-            } else {
-                throw new Error("Local close not fully supported in this block"); // Force fallback to local block below? 
-                // Actually the original code just had updateDoc for Cloud, and then a separate block for Local.
-                // So we need to ensure we don't break flow.
-                // The original code:
-                /*
-                try {
-                   await updateDoc(...);
-                   toast.success(...);
-                   return; // Returns here if cloud success
-                } catch (e) { ... }
-                
-                if (session.id...) { // Local block }
-                */
-            }
-            toast.success("Session closed and analyzed.");
-            return;
-        } catch (e) {
-            console.warn("Cloud close failed, trying local", e);
-        }
-
-        if (isDemoSession(session)) {
-            try {
-                updateLocalSession(session.id!, { status: "CLOSED" });
-                window.location.reload();
-            } catch (e) {
-                console.error(e);
-            }
-        }
-        setAnalyzing(false);
-    };
-
-    const getAggregatedData = (question: Question) => {
-        if (question.type === "multiple_choice") {
-            const counts: Record<string, number> = {};
-            question.options?.forEach(o => counts[o] = 0);
-            responses.forEach(r => {
-                if (!r.answers) return;
-                const ans = r.answers[question.id];
-                if (ans && counts[ans] !== undefined) counts[ans]++;
-            });
-            return Object.entries(counts).map(([name, value]) => ({ name, value }));
-        }
-        return [];
-    };
-
-    const getTextAnswers = (questionId: string) => {
-        return responses
-            .map(r => r.answers ? r.answers[questionId] : undefined)
-            .filter(a => a && typeof a === 'string' && !a.startsWith("http"))
-            .slice(-5);
-    };
-
-    return (
-        <div className="max-w-6xl mx-auto p-6 space-y-6">
-            <header className="flex flex-col md:flex-row justify-between items-center gap-6 bg-slate-900 text-white p-6 rounded-2xl shadow-lg">
-                {/* ... existing header ... */}
-                <div className="flex flex-col md:flex-row items-center gap-6 text-center md:text-left">
-                    <Link href="/professor/dashboard" className="md:mr-4 p-2 bg-slate-800 rounded-full hover:bg-slate-700 transition-colors">
-                        <ArrowLeft className="w-5 h-5 text-slate-300" />
-                    </Link>
-                    <div className="text-center md:text-left">
-                        <div className="text-xs text-slate-400 uppercase tracking-widest font-bold">Join Code</div>
-                        <div className="text-5xl font-mono font-bold tracking-widest">{session.code}</div>
-                        {session.title && <div className="text-lg font-serif font-bold text-slate-300 mt-1">{session.title}</div>}
-                    </div>
-                    <div className="hidden md:block h-12 w-px bg-slate-700"></div>
-                    <div className="flex items-center gap-3">
-                        <div className="p-3 bg-rose-600 rounded-full animate-pulse">
-                            <Users className="w-6 h-6 text-white" />
-                        </div>
-                        <div>
-                            <div className="text-3xl font-bold font-serif">{responses.length}</div>
-                            <div className="text-xs text-slate-400">Active Students</div>
-                        </div>
-                    </div>
-                </div>
-
-
-                <div className="flex flex-col sm:flex-row gap-4 w-full md:w-auto">
-                    <Dialog open={isAddOpen} onOpenChange={setIsAddOpen}>
-                        <DialogTrigger asChild>
-                            <Button size="lg" className="w-full sm:w-auto rounded-xl px-6 font-bold text-lg bg-indigo-600 hover:bg-indigo-700 text-white border-0">
-                                <Plus className="mr-2 w-6 h-6" /> Add Question
-                            </Button>
-                        </DialogTrigger>
-                        <DialogContent>
-                            {/* ... existing dialog ... */}
-                            <div className="space-y-4">
-                                <h3 className="text-lg font-serif font-bold">Add Live Question</h3>
-                                <div className="space-y-2">
-                                    <Label>Question Text</Label>
-                                    <Input value={newQText} onChange={e => setNewQText(e.target.value)} placeholder="What do you want to ask?" />
-                                </div>
-                                <div className="space-y-2">
-                                    <Label>Type</Label>
-                                    <Select value={newQType} onValueChange={(v: any) => setNewQType(v)}>
-                                        <SelectTrigger>
-                                            <SelectValue />
-                                        </SelectTrigger>
-                                        <SelectContent>
-                                            <SelectItem value="short_text">Short Text</SelectItem>
-                                            <SelectItem value="multiple_choice">Multiple Choice</SelectItem>
-                                            <SelectItem value="rating">Star Rating</SelectItem>
-                                            <SelectItem value="file_upload">File Upload</SelectItem>
-                                        </SelectContent>
-                                    </Select>
-                                </div>
-                                {newQType === "multiple_choice" && (
-                                    <div className="space-y-2">
-                                        <Label>Options</Label>
-                                        {newQOptions.map((opt, idx) => (
-                                            <div key={idx} className="flex gap-2">
-                                                <Input
-                                                    value={opt}
-                                                    onChange={e => {
-                                                        const newO = [...newQOptions];
-                                                        newO[idx] = e.target.value;
-                                                        setNewQOptions(newO);
-                                                    }}
-                                                />
-                                                <Button variant="ghost" size="icon" onClick={() => setNewQOptions(newQOptions.filter((_, i) => i !== idx))}>
-                                                    <Trash2 className="w-4 h-4" />
-                                                </Button>
-                                            </div>
-                                        ))}
-                                        <Button variant="link" onClick={() => setNewQOptions([...newQOptions, `Option ${newQOptions.length + 1}`])}>+ Add Option</Button>
-                                    </div>
-                                )}
-                                <Button onClick={handleAddQuestion} className="w-full">Launch Question</Button>
-                            </div>
-                        </DialogContent>
-                    </Dialog>
-                    <Dialog>
-                        <DialogTrigger asChild>
-                            <Button variant="outline" size="lg" className="w-full sm:w-auto rounded-xl px-6 font-bold text-lg bg-slate-800 border-slate-700 text-slate-300 hover:bg-slate-700 hover:text-white">
-                                <QrCode className="mr-2 w-6 h-6" /> Share
-                            </Button>
-                        </DialogTrigger>
-                        <DialogContent className="sm:max-w-md">
-                            <SessionQR sessionCode={session.code} sessionId={session.id} />
-                        </DialogContent>
-                    </Dialog>
-
-                    <Button variant="destructive" size="lg" onClick={closeSession} disabled={analyzing} className="w-full sm:w-auto rounded-xl px-8 font-bold text-lg">
-                        {analyzing ? (
-                            <>
-                                <Loader2 className="mr-2 w-6 h-6 animate-spin" /> Synthesizing...
-                            </>
-                        ) : (
-                            <>
-                                <StopCircle className="mr-2 w-6 h-6" /> Close
-                            </>
-                        )}
-                    </Button>
-                </div>
-            </header>
-
-            <div className="grid grid-cols-1 gap-6">
-                {session.questions.map((q) => {
-                    const analysis = session.analysis?.[q.id];
-                    const isActive = session.activeQuestionIds
-                        ? session.activeQuestionIds.includes(q.id)
-                        : session.activeQuestionId === q.id;
-
-                    return (
-                        <Card key={q.id} className="border-slate-200 shadow-sm overflow-hidden">
-                            <CardHeader className="bg-slate-50 border-b border-slate-100 pb-3 flex flex-row justify-between items-start">
-                                <div className="space-y-1">
-                                    <div className="flex items-center gap-2">
-                                        <CardTitle className="font-serif text-xl">{q.text}</CardTitle>
-                                        <span className={`text-[10px] uppercase font-bold px-2 py-0.5 rounded-full ${isActive ? "bg-rose-100 text-rose-700 animate-pulse" : "bg-slate-100 text-slate-500"}`}>
-                                            {isActive ? "Presenting Now" : "Hidden"}
-                                        </span>
-                                    </div>
-                                </div>
-                                <div className="flex gap-2">
-
-
-                                    <Button
-                                        variant={isActive ? "destructive" : "default"}
-                                        size="sm"
-                                        onClick={() => toggleActiveQuestion(q.id)}
-                                        className={`${isActive ? "bg-rose-600 hover:bg-rose-700 text-white" : "bg-indigo-600 hover:bg-indigo-700 text-white"}`}
-                                    >
-                                        {isActive ? (
-                                            <>
-                                                <StopCircle className="w-4 h-4 mr-2" /> Stop
-                                            </>
-                                        ) : (
-                                            <>
-                                                <Play className="w-4 h-4 mr-2" /> Show Live
-                                            </>
-                                        )}
-                                    </Button>
-                                    <Button variant="ghost" size="sm" onClick={() => deleteQuestion(q.id)} className="text-slate-400 hover:text-red-600 hover:bg-red-50">
-                                        <Trash2 className="w-4 h-4" />
-                                    </Button>
-                                </div>
-                            </CardHeader>
-                            <CardContent className={`p-6 ${q.isActive === false ? "opacity-50" : ""}`}>
-                                {q.type === "multiple_choice" ? (
-                                    <div className="h-[200px] w-full flex items-center justify-center">
-                                        {getAggregatedData(q).length > 0 ? (
-                                            <ResponsiveContainer width="100%" height="100%">
-                                                <BarChart data={getAggregatedData(q)} layout="vertical" margin={{ left: 0 }}>
-                                                    <XAxis type="number" hide />
-                                                    <YAxis dataKey="name" type="category" width={150} tick={{ fontSize: 12 }} />
-                                                    <Tooltip cursor={{ fill: 'transparent' }} contentStyle={{ borderRadius: 8 }} />
-                                                    <Bar dataKey="value" fill="var(--primary)" radius={[0, 4, 4, 0]}>
-                                                        {getAggregatedData(q).map((entry, index) => (
-                                                            <Cell key={`cell-${index}`} fill={index % 2 === 0 ? '#be123c' : '#e11d48'} />
-                                                        ))}
-                                                    </Bar>
-                                                </BarChart>
-                                            </ResponsiveContainer>
-                                        ) : (
-                                            <p className="text-slate-400 italic">Waiting for responses...</p>
-                                        )}
-                                    </div>
-                                ) : q.type === "short_text" ? (
-                                    <div className="space-y-4">
-                                        {analysis && (
-                                            <motion.div
-                                                initial={{ opacity: 0, y: 10 }}
-                                                animate={{ opacity: 1, y: 0 }}
-                                                className="bg-violet-50 border border-violet-100 rounded-xl p-5 space-y-3"
-                                            >
-                                                <div className="flex items-center gap-2 text-violet-700 font-bold font-serif">
-                                                    <Bot className="w-5 h-5" /> LUNR AI Insights
-                                                </div>
-
-                                                <div className="text-slate-700 text-sm italic border-l-2 border-violet-200 pl-3">
-                                                    "{analysis.consensus}"
-                                                </div>
-
-                                                <div className="grid grid-cols-1 md:grid-cols-2 gap-4 pt-2">
-                                                    <div>
-                                                        <h4 className="text-xs font-bold text-slate-500 uppercase tracking-wider mb-2">Distribution</h4>
-                                                        <p className="text-sm text-slate-700">{analysis.distribution_analysis}</p>
-                                                    </div>
-                                                    <div>
-                                                        <h4 className="text-xs font-bold text-slate-500 uppercase tracking-wider mb-2">Key Inferences</h4>
-                                                        <ul className="list-disc list-inside text-sm text-slate-700 space-y-1">
-                                                            {analysis.key_inferences?.map((inf: string, i: number) => (
-                                                                <li key={i}>{inf}</li>
-                                                            ))}
-                                                        </ul>
-                                                    </div>
-                                                </div>
-
-
-                                                {analysis.outlier_insight && (
-                                                    <div className="bg-white p-3 rounded-lg border border-violet-100 text-sm text-slate-600">
-                                                        <span className="font-bold text-violet-600">Outlier:</span> {analysis.outlier_insight}
-                                                    </div>
-                                                )}
-                                            </motion.div>
-                                        )}
-
-                                        <div className="space-y-2">
-                                            <AnimatePresence>
-                                                {getTextAnswers(q.id).map((ans, i) => (
-                                                    <motion.div
-                                                        key={i}
-                                                        initial={{ opacity: 0, x: -10 }}
-                                                        animate={{ opacity: 1, x: 0 }}
-                                                        className={`p-3 rounded-lg border text-lg font-medium ${i === 0 ? "bg-white border-primary/20 shadow-sm text-slate-800" : "bg-slate-50 border-transparent text-slate-500"
-                                                            }`}
-                                                    >
-                                                        &quot;{ans}&quot;
-                                                    </motion.div>
-                                                ))}
-                                            </AnimatePresence>
-                                            {getTextAnswers(q.id).length === 0 && <p className="text-slate-400 italic">No responses yet...</p>}
-                                        </div>
-                                    </div>
-                                ) : q.type === "rating" ? (
-                                    <div className="flex flex-col md:flex-row gap-8 items-center justify-center py-8">
-                                        {(() => {
-                                            const answers = responses
-                                                .map(r => r.answers ? parseFloat(r.answers[q.id] || "0") : 0)
-                                                .filter(v => v > 0);
-
-                                            const avg = answers.length > 0 ? (answers.reduce((a, b) => a + b, 0) / answers.length) : 0;
-
-                                            // Histogram
-                                            const counts: Record<string, number> = { "5 Stars": 0, "4 Stars": 0, "3 Stars": 0, "2 Stars": 0, "1 Star": 0 };
-                                            answers.forEach(a => {
-                                                const bucket = Math.round(a); // Round 4.5 to 5 for histogram bucket
-                                                if (bucket >= 1 && bucket <= 5) counts[`${bucket} ${bucket === 1 ? "Star" : "Stars"}`]++;
-                                            });
-                                            const chartData = Object.entries(counts).map(([name, value]) => ({ name, value }));
-
-                                            return (
-                                                <>
-                                                    <div className="flex flex-col items-center">
-                                                        <div className="text-6xl font-serif font-bold text-slate-800">{avg.toFixed(1)}</div>
-                                                        <div className="mb-2">
-                                                            <StarRating value={avg} readOnly size="lg" />
-                                                        </div>
-                                                        <div className="text-slate-400 text-sm">{answers.length} Responses</div>
-                                                    </div>
-
-                                                    {answers.length > 0 && (
-                                                        <div className="h-[200px] w-full md:w-1/2">
-                                                            <ResponsiveContainer width="100%" height="100%">
-                                                                <BarChart data={chartData} layout="vertical" margin={{ left: 0 }}>
-                                                                    <XAxis type="number" hide />
-                                                                    <YAxis dataKey="name" type="category" width={80} tick={{ fontSize: 12 }} />
-                                                                    <Tooltip cursor={{ fill: 'transparent' }} contentStyle={{ borderRadius: 8 }} />
-                                                                    <Bar dataKey="value" fill="var(--primary)" radius={[0, 4, 4, 0]}>
-                                                                        {chartData.map((entry, index) => (
-                                                                            <Cell key={`cell-${index}`} fill='#fbbf24' />
-                                                                        ))}
-                                                                    </Bar>
-                                                                </BarChart>
-                                                            </ResponsiveContainer>
-                                                        </div>
-                                                    )}
-                                                </>
-                                            );
-                                        })()}
-                                    </div>
-                                ) : (
-                                    <div className="text-slate-500 italic">File uploads are collected in storage. Visualization coming soon.</div>
-                                )}
-                            </CardContent>
-                        </Card>
-                    );
-                })}
-            </div>
-        </div>
-    );
-}
