+++ src/components/professor/LiveDashboard/index.ts
@@ -1,0 +1,1 @@
+export { LiveDashboard } from "./LiveDashboard";
